+++ src/main.rs
@@ -1,12 +1,6 @@
-mod ast;
-mod ast_helpers;
-mod inference;
-mod pest_parser;
-mod substitution;
-mod r#type;
-mod type_environment;
-mod type_error;
-mod typer;
+mod syntax;
+mod typesystem;
+
 fn main() {
     println!("Hello, world!");
 }
