+++ src/typesystem/mod.rs
@@ -1,0 +1,6 @@
+mod inference;
+mod substitution;
+pub mod r#type;
+mod type_environment;
+pub mod type_error;
+pub mod typer;
