+++ src/typesystem/type.rs
@@ -1,7 +1,6 @@
+use crate::typesystem::substitution::{Substitutable, Substitution};
 use std::collections::HashSet;
 use std::fmt::{Debug, Formatter};
-
-use crate::substitution::{Substitutable, Substitution};
 
 #[derive(Clone, PartialEq, Eq)]
 pub enum Type {
