+++ src/typesystem/typer.rs
@@ -1,12 +1,12 @@
 use std::cell::Cell;
 use std::collections::HashMap;
 
-use crate::ast::{ArgList, Expression, Param, ParamList, TypeParam};
-use crate::inference::Inference;
-use crate::r#type::{PrimitiveType, Type, TypeScheme};
-use crate::substitution::{Substitutable, Substitution};
-use crate::type_environment::TypeEnvironment;
-use crate::type_error::{TypeError, TypeErrorReport};
+use crate::syntax::ast::{ArgList, Expression, Param, ParamList, TypeParam};
+use crate::typesystem::inference::Inference;
+use crate::typesystem::r#type::{PrimitiveType, Type, TypeScheme};
+use crate::typesystem::substitution::{Substitutable, Substitution};
+use crate::typesystem::type_environment::TypeEnvironment;
+use crate::typesystem::type_error::{TypeError, TypeErrorReport};
 
 type InferenceResult = Result<Inference, TypeErrorReport>;
 
@@ -270,10 +270,10 @@ impl<'a> Typer<'a> {
 
 #[cfg(test)]
 mod tests {
-    use crate::ast::{ArgList, Expression, Param, ParamList, TypeParam};
-    use crate::r#type::{PrimitiveType, Type};
-    use crate::type_error::TypeError;
-    use crate::typer::{TypeEnvironment, Typer};
+    use crate::syntax::ast::{ArgList, Expression, Param, ParamList, TypeParam};
+    use crate::typesystem::r#type::{PrimitiveType, Type};
+    use crate::typesystem::type_error::TypeError;
+    use crate::typesystem::typer::{TypeEnvironment, Typer};
     use crate::{
         application, bool_lit, infix, int_lit, lambda, let_expr, named, param, string_lit,
         type_param,
